requirements.txt
einops==0.4.1
matplotlib==3.5.1
numpy==1.21.2
opencv_python==4.5.5.64
Pillow==9.3.0
timm==0.4.5
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
open-clip-torch
wandb
